x/interchainstaking: keeper, types: fix non-determinism from map iteration
In removing stale delegations, the loop was non-deterministic due to
iterating over a map. This change fixes that by extracting the keys
firstly into a slice, sorting it then iterating using those.

Also while here fixed non-determinism in types.Error which produced
different results on each invocation of types.Error.Error() and added
a unit test to lock this behavior in and avoid regressions.

Fixes #140