x/participationrewards/keeper: move cheap prefix checks to front + better error UX
This change moves the cheaper prefix checks per request to the top of
each of UmeeTotalBorrowsUpdateCallback and UmeeInterestScalarUpdateCallback,
prefix checks as it is unnecessary to invoke an expensive unmarshalling
yet the result will be discarded from the result of a simple equals
check. While here, also improved the error user experience (UX) for prefix
mismatches to return the prefix that was retrieved vs expected.

Fixes #776