fix(icq-relayer/config): return immediately on non-os.IsNotExist errors
This change avoids the need to try to brute-force creating
an entirely fresh configuration, in the case where an error
that isn't a path non-existence error occurs, for example
in the case of a permission error on the parent directory
or non-matching user/group permissions. Now we check against
the returned error after os.Stat and accordingly return
immediately or continue to create the configuration.

Fixes #1524